paraphrase/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
filelock==3.12.0
huggingface-hub==0.13.4
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
lit==16.0.1
Markdown==3.4.3
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pegasus==0.1.3
protobuf==3.20.3
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
sacremoses==0.0.53
sentence-splitter==1.4
sentencepiece==0.1.98
six==1.16.0
sqlparse==0.4.4
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.28.1
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
